# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Sendpaper API",
    "version": "1.0.0",
    "description": "Print and mail postcards and letters to US addresses. Every order returns a checkout_url a person pays; nothing is printed until then. No API key required."
  },
  "servers": [
    {
      "url": "https://sendmypaper.com/v1"
    }
  ],
  "paths": {
...
    "/orders/{id}": {
      "get": {
        "summary": "Retrieve an order",
        "operationId": "getOrder",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "ord_8k2m4qz7c1x9b3vd"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
...
    "responses": {
...
      "NotFound": {
        "description": "No such order",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      }
    },
    "schemas": {
      "Address": {
        "type": "object",
        "required": [
          "name",
          "line1",
          "city",
          "state",
          "zip"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 60
          },
          "company": {
            "type": "string",
            "maxLength": 60
          },
          "line1": {
            "type": "string",
            "maxLength": 64,
            "description": "Street address"
          },
          "line2": {
            "type": "string",
            "maxLength": 64,
            "description": "Apartment, suite, unit"
          },
          "city": {
            "type": "string",
            "maxLength": 40
          },
          "state": {
            "type": "string",
            "description": "Two-letter US state or territory code",
            "example": "CA"
          },
          "zip": {
            "type": "string",
            "pattern": "^\\d{5}(-\\d{4})?$",
            "example": "94105"
          }
        }
      },
...
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "ord_8k2m4qz7c1x9b3vd"
          },
          "object": {
            "type": "string",
            "enum": [
              "postcard",
              "letter"
            ]
          },
          "product": {
            "type": "string",
            "enum": [
              "postcard_4x6",
              "postcard_6x9",
              "letter"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "awaiting_payment",
              "paid",
              "printing",
              "mailed",
              "cancelled",
              "refunded"
            ]
          },
          "status_detail": {
            "type": "string"
          },
          "price": {
            "type": "object",
            "properties": {
              "amount_cents": {
                "type": "integer"
              },
              "currency": {
                "type": "string"
              },
              "display": {
                "type": "string"
              }
            }
          },
          "to": {
            "$ref": "#/components/schemas/Address"
          },
          "from": {
            "$ref": "#/components/schemas/Address"
          },
          "content": {
            "type": "object"
          },
          "checkout_url": {
            "type": [
              "string",
              "null"
            ],
            "description": "Where the person pays. Null once paid or closed. Never expires."
          },
          "order_url": {
            "type": "string",
            "description": "Human-readable status page"
          },
          "preview_url": {
            "type": "string",
            "description": "Exact print preview"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "paid_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "mailed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "events": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "at": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "note": {
                  "type": "string"
                }
              }
            }
          },
          "certified": {
            "type": "string",
            "enum": [
              "none",
              "certified",
              "certified_return_receipt"
            ],
            "description": "Letters only: the USPS extra service."
          },
          "tracking": {
            "type": [
              "object",
              "null"
            ],
            "description": "USPS tracking once the piece is accepted (certified letters always get one).",
            "properties": {
              "number": {
                "type": "string"
              },
              "url": {
                "type": "string",
                "format": "uri"
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "example": "invalid_request"
              },
              "message": {
                "type": "string"
              },
              "fields": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "field": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}